Replace generate_random_username with generate_unique_username
Do not provide a default username on the User model
as there are few instances where User objects are
created and we want to be able to call
`generate_unique_username` which queries the User
model. `generate_random_username` results in too
many username collisions when creating lots of users.